#3 Move utility function dag_from_strings() out of dag_test and into dag module
