+++ src/tests/dag_test.py
@@ -7,6 +7,7 @@ from parse2003 import ParseError
 import make_dag
 from dag_node import DAGError
 from fparser import Fortran2003
+from dag import dag_from_strings
 
 # constants
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
@@ -36,35 +37,6 @@ class Options(object):
         self.show_weights = False
         self.unroll_factor = 1
         self.mode = 'auto'
-
-
-def dag_from_strings(lines, name=None):
-    ''' Function that takes a list of strings (containing Fortran
-    assignment statements) and generates a DAG '''
-    from dag import DirectedAcyclicGraph
-    from parse2003 import Variable
-    assigns = []
-    for line in lines:
-        assigns.append(Fortran2003.Assignment_Stmt(line))
-    mapping = {}
-    if name:
-        dag_name = name
-    else:
-        dag_name = "Test dag"
-
-    dag = DirectedAcyclicGraph(dag_name)
-
-    for assign in assigns:
-        lhs_var = Variable()
-        lhs_var.load(assign.items[0], mapping=mapping, lhs=True)
-        lhs_node = dag.get_node(parent=None,
-                                variable=lhs_var)
-        if assign.items[0] in mapping:
-            mapping[assign.items[0]] += "'"
-        else:
-            mapping[assign.items[0]] = assign.items[0]
-        dag.make_dag(lhs_node, assign.items[2:], mapping)
-    return dag
 
 # The tests themselves...
